src/requirements.txt
dlib
sklearn
opencv-python
numpys